remove misleading content from index.html
index.html used to be used for development purposes when we didn't run a clojure server route via lein run.
now that we don't use this file it can present as 'loading' interminably which can confused devs new to bluegenes who have accidentally gone to the port in the figwheel prompt.
I've removed all content from it along with a link telling users to go to the correct port - usually 5000 - and updated the comments in the files accordingly.