Upgrade to React 16 and update dependencies
Something to note is that react-with-addons has been replaced by react
as included by reagent, and react-transition-group has been included
through cljsjs. This is because it has been separated from react.

I also removed highlight.js from npm-deps as it was causing a conflict
with the cljsjs dependencies of re-frame-10x 0.4.x. It seems like
highlight.js is only used in im-tables. In the future, it would be
reasonable to switch to using react-highlight through cljsjs instead.