Fix tools loading then disappearing on list page
This would happen if you opened a report page, then opened the results
for a list later on. The tools entity from the report page would still
be present, so the list page would load and show the tools for the
previous report page, before the tool entity got updated. This caused a
huge slowdown when a lot of tools were present.