Report: Remove customizable default layout
It only made sense in the context of using referencedType. Proteins have
a Canonical Protein and Isoforms, both of which use the Protein
referencedType, causing them not to be distinguishable in the layout
editor. A previous commit migrated to using the name property instead,
which won't work with a generic default layout.
It doesn't seem like the original vision of the default layout was
achievable (the fallback layout handles this better), hence the decision
to remove the default layout.