Detect expired token more aggressively and improve alert text
Until now, an expired token could be missed as it was up to each
individual :on-failure event to handle it. Turns out this isn't a good
idea, and :on-unauthorised will instead be required to override the
dispatching of :flag-invalid-token. The downside of this is 401 could
have a different meaning in other contexts (like not being authorized to
delete a public list) and these will now need an :on-unauthorised event
to avoid the "expired token" alert appearing (which would be confusing).