Fix OAuth2 for Firefox and other backend improvements
The HTML having a meta tag with utf-8 charset, while the response not
including this charset in Content-Type, caused Firefox to request the
HTML twice, making session.init not work.
I noticed that wrap-restful-format caused all Content-Type to be
application/json. I fixed this by replacing it with muuntaja's
wrap-format and wrap-params.
When reading the Ring documentation, I learned that cookies should
be recreated after privilege escalation.
Finally, I added performance and security improvements to wrap-reload
and wrap-session.