Include vega graphing library using CDN instead of bundling
The main reason for this was that cljsjs/vega-lite had too big version
jumps. The last release has a bug with labelOrient=bottom not working,
and the release before doesn't render the piechart. This switches to
using the CDN instead, which is probably a better approach anyways, as
it doesn't add to the Bluegenes bundle size and will be cached across
Bluegenes instances.