Add new optional :bluegenes-backend-service-root config var
This is to be used when you want the BG backend to use a different
address to reach the default service (tomcat server running InterMine)
instead of the address the BG frontend uses. Usually this won't be
necessary as :bluegenes-default-service-root should be reachable even
from the backend, but this might not be the case in test deployments.
Using the internal route also results in better performance, although
there are not many requests the backend makes to the IM service.