Support portal.do linking in to upload and don't reset
Dispatching idresolver.events/reset from upload.views/main would send
the user back to the input step when linking in. I decided to remove
this as it also seems better to persist the input data until the user
decides to clear it, instead of resetting it whenever the user opens the
page.